feat(DTFS2-7298): make facility mapping file structure consistent between GEF and non-GEF facilities
