+++ trade-finance-manager-api/src/v1/rest-mappings/facility.js
@@ -1,22 +1,19 @@
 const mapFacilitySnapshot = require('./mappings/facilities/mapFacilitySnapshot');
 const mapFacilityTfm = require('./mappings/facilities/mapFacilityTfm');
-const mapGefFacility = require('./mappings/gef-facilities/mapGefFacility');
+const mapGefFacilitySnapshot = require('./mappings/gef-facilities/mapGefFacilitySnapshot');
 const isGefFacility = require('./helpers/isGefFacility');
 
 const facilityMapper = (facility, dealSnapshot, dealTfm) => {
   const facilityType = facility.facilitySnapshot.type;
+  const isGef = isGefFacility(facilityType);
 
-  if (isGefFacility(facilityType)) {
-    return mapGefFacility(facility, dealSnapshot, dealTfm);
-  }
+  const mappedFacilitySnapshot = isGef ? mapGefFacilitySnapshot(facility, dealSnapshot) : mapFacilitySnapshot(facility, dealSnapshot.details);
 
-  const result = {
+  return {
     _id: facility._id,
-    facilitySnapshot: mapFacilitySnapshot(facility, dealSnapshot.details),
+    facilitySnapshot: mappedFacilitySnapshot,
     tfm: mapFacilityTfm(facility.tfm, dealTfm, facility),
   };
-
-  return result;
 };
 
 module.exports = facilityMapper;
