+++ trade-finance-manager-api/src/v1/rest-mappings/mappings/deal/mapDealSnapshot.js
@@ -12,7 +12,7 @@ const mapDealSnapshot = (deal) => {
     ...dealSnapshot,
     submissionDetails: mapSubmissionDetails(submissionDetails),
     eligibility: eligibility ? mapEligibility(eligibility) : {},
-    facilities: mapFacilities(facilities, dealSnapshot.details, dealTfm),
+    facilities: mapFacilities(facilities, dealSnapshot, dealTfm),
     totals: mapTotals(facilities),
     isFinanceIncreasing: false,
   };
