+++ trade-finance-manager-api/src/v1/rest-mappings/mappings/facilities/mapFacilities.js
@@ -1,16 +1,9 @@
-const mapFacilitySnapshot = require('./mapFacilitySnapshot');
-const mapFacilityTfm = require('./mapFacilityTfm');
+const facilityMapper = require('../../facility');
 
-const mapFacilities = (facilities, dealDetails, dealTfm) => {
+const mapFacilities = (facilities, dealSnapshot, dealTfm) => {
   // Map facilities if only they exists
   if (facilities?.length) {
-    const mappedFacilities = facilities.map((facility) => ({
-      _id: facility._id,
-      facilitySnapshot: mapFacilitySnapshot(facility, dealDetails),
-      tfm: mapFacilityTfm(facility.tfm, dealTfm, facility),
-    }));
-
-    return mappedFacilities;
+    return facilities.map((facility) => facilityMapper(facility, dealSnapshot, dealTfm));
   }
 
   return null;
