+++ trade-finance-manager-api/src/v1/rest-mappings/mappings/gef-facilities/mapGefFacilities.js
@@ -1,13 +1,11 @@
-const mapGefFacility = require('./mapGefFacility');
+const facilityMapper = require('../../facility');
 
 const mapGefFacilities = (dealSnapshot, dealTfm) => {
   const { facilities } = dealSnapshot;
 
   // Map facilities if only they exists
   if (facilities?.length) {
-    const mappedFacilities = facilities.map((facility) => mapGefFacility(facility, dealSnapshot, dealTfm));
-
-    return mappedFacilities;
+    return facilities.map((facility) => facilityMapper(facility, dealSnapshot, dealTfm));
   }
 
   return null;
