+++ trade-finance-manager-api/src/v1/rest-mappings/mappings/gef-facilities/mapGefFacilitySnapshot.js
@@ -4,7 +4,6 @@ const mapFacilityValue = require('../facilities/mapFacilityValue');
 const mapFacilityProduct = require('../facilities/mapFacilityProduct');
 const mapFacilityType = require('../facilities/mapFacilityType');
 const mapGuaranteeFeePayableToUkef = require('../facilities/mapGuaranteeFeePayableToUkef');
-const mapFacilityTfm = require('../facilities/mapFacilityTfm');
 const mapGefUkefFacilityType = require('./mapGefUkefFacilityType');
 const mapGefFacilityDates = require('./mapGefFacilityDates');
 const mapFacilityValueExportCurrency = require('../facilities/mapFacilityValueExportCurrency');
@@ -15,7 +14,7 @@ const mapUkefExposureValue = require('../facilities/mapUkefExposureValue');
  * Note: This implimentation modifies the facility snapshot
  * to have values not consistent with the facility snapshot in the database.
  */
-const mapGefFacility = (facility, dealSnapshot, dealTfm) => {
+const mapGefFacilitySnapshot = (facility, dealSnapshot) => {
   const { facilitySnapshot, tfm: facilityTfm } = facility;
 
   const { dealId, coverPercentage, currency, value, interestPercentage, feeType, feeFrequency, hasBeenIssued, name, type, ukefFacilityId, guaranteeFee } =
@@ -29,39 +28,33 @@ const mapGefFacility = (facility, dealSnapshot, dealTfm) => {
 
   facilitySnapshot.ukefFacilityType = type;
 
-  const result = {
+  return {
     _id: facility._id,
-    facilitySnapshot: {
-      _id: facility._id,
-      isGef: true,
-      dealId,
-      bankFacilityReference: name,
-      banksInterestMargin: `${interestPercentage}%`,
-      coveredPercentage: `${coverPercentage}%`,
-      dates: mapGefFacilityDates(facility, facilityTfm, dealSnapshot),
-      facilityProduct: facilitySnapshot.facilityProduct,
-      facilityStage: facilitySnapshot.facilityStage,
-      hasBeenIssued: facilitySnapshot.hasBeenIssued,
-      type: mapFacilityType(facilitySnapshot),
-      currency: currency.id,
-      facilityValueExportCurrency: mapFacilityValueExportCurrency(facility),
-      value: mapFacilityValue(currency.id, formattedFacilityValue, facility),
-      feeType,
-      feeFrequency,
-      guaranteeFeePayableToUkef: mapGuaranteeFeePayableToUkef(guaranteeFee),
-      dayCountBasis: facilitySnapshot.dayCountBasis,
+    isGef: true,
+    dealId,
+    bankFacilityReference: name,
+    banksInterestMargin: `${interestPercentage}%`,
+    coveredPercentage: `${coverPercentage}%`,
+    dates: mapGefFacilityDates(facility, facilityTfm, dealSnapshot),
+    facilityProduct: facilitySnapshot.facilityProduct,
+    facilityStage: facilitySnapshot.facilityStage,
+    hasBeenIssued: facilitySnapshot.hasBeenIssued,
+    type: mapFacilityType(facilitySnapshot),
+    currency: currency.id,
+    facilityValueExportCurrency: mapFacilityValueExportCurrency(facility),
+    value: mapFacilityValue(currency.id, formattedFacilityValue, facility),
+    feeType,
+    feeFrequency,
+    guaranteeFeePayableToUkef: mapGuaranteeFeePayableToUkef(guaranteeFee),
+    dayCountBasis: facilitySnapshot.dayCountBasis,
 
-      // TODO: DTFS2-4634 - we shouldn't need type and ukefFacilityType.
-      ukefFacilityType: mapGefUkefFacilityType(type),
-      ukefFacilityId,
-      ukefExposure: mapUkefExposureValue(facilityTfm, facility),
-      providedOn: facilitySnapshot.details,
-      providedOnOther: facilitySnapshot.detailsOther,
-    },
-    tfm: mapFacilityTfm(facilityTfm, dealTfm, facility),
+    // TODO: DTFS2-4634 - we shouldn't need type and ukefFacilityType.
+    ukefFacilityType: mapGefUkefFacilityType(type),
+    ukefFacilityId,
+    ukefExposure: mapUkefExposureValue(facilityTfm, facility),
+    providedOn: facilitySnapshot.details,
+    providedOnOther: facilitySnapshot.detailsOther,
   };
-
-  return result;
 };
 
-module.exports = mapGefFacility;
+module.exports = mapGefFacilitySnapshot;
